split grid world
