+++ scripts/grid_intake_horizon_experiment.py
@@ -9,9 +9,8 @@ import numpy as np
 import pandas as pd
 from matplotlib import pyplot as plt
 
-from multitask_personalization.envs.grid_world.grid_world import (
-    _EMPTY,
-    _OBSTACLE,
+from multitask_personalization.envs.grid_world.grid_world_mdp import _EMPTY, _OBSTACLE
+from multitask_personalization.envs.grid_world.grid_world_tasks import (
     GridTask,
 )
 from multitask_personalization.methods.approach import Approach
